+++ src/main/java/org/ticket/exception/EventNotFoundException.java
@@ -2,7 +2,7 @@ package org.ticket.exception;
 
 public class EventNotFoundException extends RuntimeException {
 
-    public EventNotFoundException(Integer eventId) {
+    public EventNotFoundException(Long eventId) {
         super("No event with id " + eventId);
     }
 }
